+++ packages/adapter-commons/src/index.ts
@@ -1,20 +1,20 @@
-const { _ } = require('@feathersjs/commons');
+import { _ } from '@feathersjs/commons';
 
-const AdapterService = require('./service');
-const filterQuery = require('./filter-query');
-const sort = require('./sort');
+export { AdapterService, InternalServiceMethods, ServiceOptions } from './service';
+export { default as filterQuery, FILTERS, OPERATORS } from './filter-query';
+export * from './sort';
 
 // Return a function that filters a result object or array
 // and picks only the fields passed as `params.query.$select`
 // and additional `otherFields`
-const select = function select (params, ...otherFields) {
+export function select (params: any, ...otherFields: any[]) {
   const fields = params && params.query && params.query.$select;
 
   if (Array.isArray(fields) && otherFields.length) {
     fields.push(...otherFields);
   }
 
-  const convert = result => {
+  const convert = (result: any) => {
     if (!Array.isArray(fields)) {
       return result;
     }
@@ -22,17 +22,11 @@ const select = function select (params, ...otherFields) {
     return _.pick(result, ...fields);
   };
 
-  return result => {
+  return (result: any) => {
     if (Array.isArray(result)) {
       return result.map(convert);
     }
 
     return convert(result);
   };
-};
-
-module.exports = Object.assign({
-  select,
-  filterQuery,
-  AdapterService
-}, sort);
+}